+++ gxmetal/guest/apps/GXMetalTest.r
@@ -27,9 +27,9 @@ resource 'BNDL' (128, purgeable) {
 };
 
 resource 'vers' (1) {
-    2, 0, release, 3, verUS,
-    "2.0.3",
-    "GXMetal Test 2.0.3"
+    2, 0, release, 4, verUS,
+    "2.0.4",
+    "GXMetal Test 2.0.4"
 };
 
 resource 'cfrg' (0) {