+++ gxmetal/guest/src/GXMetal.r
@@ -19,9 +19,9 @@ data 'ftag' (0) {
 };
 
 resource 'vers' (1) {
-    2, 0, release, 3, verUS,
-    "2.0.3",
-    "GXMetal 2.0.3"
+    2, 0, release, 4, verUS,
+    "2.0.4",
+    "GXMetal 2.0.4"
 };
 
 resource 'cfrg' (0) {