+++ gxmetal/guest/src/GXMetalInput.r
@@ -5,9 +5,9 @@
 #include "GXMetalIcon.r"
 
 resource 'vers' (1) {
-    2, 0, release, 3, verUS,
-    "2.0.3",
-    "GXMetal Input 2.0.3"
+    2, 0, release, 4, verUS,
+    "2.0.4",
+    "GXMetal Input 2.0.4"
 };
 
 resource 'cfrg' (0) {